app/requirements.txt
streamlit
streamlit-chat
google-cloud-aiplatform
pandas
openpyxl
annoy
requests
requests-mock
python-dotenv
sqlalchemy
psycopg2
langchain-core
langchain-google-vertexai
fluent-logger
ddtrace>=2.7.3
datadog
alembic
google-cloud-discoveryengine
Flask
google-cloud-storage
markdownify
guardrails-ai
gunicorn
celery
flower
redis
flask-cors
google-auth
trino
time-machine
langchain-google-alloydb-pg
langchain-google-vertexai
jwt
redis